# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Question Bank API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Subject {
  id: number;
  name: string;
  createdAt: string;
  /** @nullable */
  chapterCount?: number | null;
  /** @nullable */
  questionCount?: number | null;
}

export interface CreateSubjectBody {
  name: string;
}

export interface Chapter {
  id: number;
  subjectId: number;
  /** @nullable */
  subjectName?: string | null;
  name: string;
  createdAt: string;
  /** @nullable */
  questionCount?: number | null;
}

export interface CreateChapterBody {
  subjectId: number;
  name: string;
}

export type QuestionType = (typeof QuestionType)[keyof typeof QuestionType];

export const QuestionType = {
  MCQ: "MCQ",
  FILLUP: "FILLUP",
} as const;

export type QuestionDifficulty =
  (typeof QuestionDifficulty)[keyof typeof QuestionDifficulty];

export const QuestionDifficulty = {
  EASY: "EASY",
  MEDIUM: "MEDIUM",
  HARD: "HARD",
} as const;

export interface Question {
  id: number;
  chapterId: number;
  /** @nullable */
  chapterName?: string | null;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  text: string;
  type: QuestionType;
  difficulty: QuestionDifficulty;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageName?: string | null;
  /** @nullable */
  imageType?: string | null;
  createdAt: string;
  /** @nullable */
  choiceCount?: number | null;
}

export interface QuestionListResponse {
  questions: Question[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateQuestionBodyType =
  (typeof CreateQuestionBodyType)[keyof typeof CreateQuestionBodyType];

export const CreateQuestionBodyType = {
  MCQ: "MCQ",
  FILLUP: "FILLUP",
} as const;

export type CreateQuestionBodyDifficulty =
  (typeof CreateQuestionBodyDifficulty)[keyof typeof CreateQuestionBodyDifficulty];

export const CreateQuestionBodyDifficulty = {
  EASY: "EASY",
  MEDIUM: "MEDIUM",
  HARD: "HARD",
} as const;

export interface CreateQuestionBody {
  chapterId: number;
  text: string;
  type: CreateQuestionBodyType;
  difficulty: CreateQuestionBodyDifficulty;
  image?: Blob;
}

export type UpdateQuestionBodyType =
  (typeof UpdateQuestionBodyType)[keyof typeof UpdateQuestionBodyType];

export const UpdateQuestionBodyType = {
  MCQ: "MCQ",
  FILLUP: "FILLUP",
} as const;

export type UpdateQuestionBodyDifficulty =
  (typeof UpdateQuestionBodyDifficulty)[keyof typeof UpdateQuestionBodyDifficulty];

export const UpdateQuestionBodyDifficulty = {
  EASY: "EASY",
  MEDIUM: "MEDIUM",
  HARD: "HARD",
} as const;

export type UpdateQuestionBodyRemoveImage =
  (typeof UpdateQuestionBodyRemoveImage)[keyof typeof UpdateQuestionBodyRemoveImage];

export const UpdateQuestionBodyRemoveImage = {
  true: "true",
  false: "false",
} as const;

export interface UpdateQuestionBody {
  chapterId?: number;
  text?: string;
  type?: UpdateQuestionBodyType;
  difficulty?: UpdateQuestionBodyDifficulty;
  image?: Blob;
  removeImage?: UpdateQuestionBodyRemoveImage;
}

export interface Choice {
  id: number;
  questionId: number;
  text: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageName?: string | null;
  isCorrect: boolean;
  createdAt: string;
}

export interface CreateChoiceBody {
  questionId: number;
  text: string;
  isCorrect: boolean;
  image?: Blob;
}

export type UpdateChoiceBodyRemoveImage =
  (typeof UpdateChoiceBodyRemoveImage)[keyof typeof UpdateChoiceBodyRemoveImage];

export const UpdateChoiceBodyRemoveImage = {
  true: "true",
  false: "false",
} as const;

export interface UpdateChoiceBody {
  text?: string;
  isCorrect?: boolean;
  image?: Blob;
  removeImage?: UpdateChoiceBodyRemoveImage;
}

export type QuestionDetailType =
  (typeof QuestionDetailType)[keyof typeof QuestionDetailType];

export const QuestionDetailType = {
  MCQ: "MCQ",
  FILLUP: "FILLUP",
} as const;

export type QuestionDetailDifficulty =
  (typeof QuestionDetailDifficulty)[keyof typeof QuestionDetailDifficulty];

export const QuestionDetailDifficulty = {
  EASY: "EASY",
  MEDIUM: "MEDIUM",
  HARD: "HARD",
} as const;

export interface QuestionDetail {
  id: number;
  chapterId: number;
  /** @nullable */
  chapterName?: string | null;
  /** @nullable */
  subjectId?: number | null;
  /** @nullable */
  subjectName?: string | null;
  text: string;
  type: QuestionDetailType;
  difficulty: QuestionDetailDifficulty;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  imageName?: string | null;
  /** @nullable */
  imageType?: string | null;
  createdAt: string;
  choices: Choice[];
}

export interface ChoicePreview {
  id: number;
  text: string;
  isCorrect: boolean;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  imageType?: string | null;
}

export type QuestionPreviewType =
  (typeof QuestionPreviewType)[keyof typeof QuestionPreviewType];

export const QuestionPreviewType = {
  MCQ: "MCQ",
  FILLUP: "FILLUP",
} as const;

export type QuestionPreviewDifficulty =
  (typeof QuestionPreviewDifficulty)[keyof typeof QuestionPreviewDifficulty];

export const QuestionPreviewDifficulty = {
  EASY: "EASY",
  MEDIUM: "MEDIUM",
  HARD: "HARD",
} as const;

export interface QuestionPreview {
  id: number;
  text: string;
  type: QuestionPreviewType;
  difficulty: QuestionPreviewDifficulty;
  /** @nullable */
  chapterName?: string | null;
  /** @nullable */
  subjectName?: string | null;
  /** @nullable */
  imageData?: string | null;
  /** @nullable */
  imageType?: string | null;
  choices: ChoicePreview[];
}

export interface DashboardStats {
  totalSubjects: number;
  totalChapters: number;
  totalQuestions: number;
  totalMCQ: number;
  totalFillup: number;
  easyCount: number;
  mediumCount: number;
  hardCount: number;
}

export interface SubjectBreakdown {
  subjectId: number;
  subjectName: string;
  questionCount: number;
  chapterCount: number;
}

export interface DifficultyBreakdown {
  difficulty: string;
  count: number;
}

export interface ExportSelectedBody {
  questionIds: number[];
  title?: string;
}

export type ListSubjectsParams = {
  search?: string;
};

export type ListChaptersParams = {
  subjectId?: number;
  search?: string;
};

export type ListQuestionsParams = {
  chapterId?: number;
  subjectId?: number;
  search?: string;
  difficulty?: ListQuestionsDifficulty;
  type?: ListQuestionsType;
  page?: number;
  limit?: number;
};

export type ListQuestionsDifficulty =
  (typeof ListQuestionsDifficulty)[keyof typeof ListQuestionsDifficulty];

export const ListQuestionsDifficulty = {
  EASY: "EASY",
  MEDIUM: "MEDIUM",
  HARD: "HARD",
} as const;

export type ListQuestionsType =
  (typeof ListQuestionsType)[keyof typeof ListQuestionsType];

export const ListQuestionsType = {
  MCQ: "MCQ",
  FILLUP: "FILLUP",
} as const;