# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Question Bank API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chapters:
    get:
      operationId: listChapters
      tags: [chapters]
      summary: List chapters, optionally filtered by subject
      parameters:
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of chapters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chapter"
    post:
      operationId: createChapter
      tags: [chapters]
      summary: Create a chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChapterBody"
      responses:
        "201":
          description: Created chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /chapters/{id}:
    get:
      operationId: getChapter
      tags: [chapters]
      summary: Get a chapter by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateChapter
      tags: [chapters]
      summary: Update a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChapterBody"
      responses:
        "200":
          description: Updated chapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chapter"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteChapter
      tags: [chapters]
      summary: Delete a chapter
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # QUESTIONS
...
  /choices:
    post:
      operationId: createChoice
      tags: [choices]
      summary: Create a choice for a question
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateChoiceBody"
      responses:
        "201":
          description: Created choice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Choice"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /choices/{id}:
    put:
      operationId: updateChoice
      tags: [choices]
      summary: Update a choice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateChoiceBody"
      responses:
        "200":
          description: Updated choice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Choice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteChoice
      tags: [choices]
      summary: Delete a choice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /choices/{id}/image:
    get:
      operationId: getChoiceImage
      tags: [choices]
      summary: Serve choice image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Image binary
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: No image

  # DASHBOARD
...
  /export/pdf/question/{id}:
    get:
      operationId: exportQuestionPdf
      tags: [export]
      summary: Export single question to PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
  /export/pdf/chapter/{id}:
    get:
      operationId: exportChapterPdf
      tags: [export]
      summary: Export all questions in a chapter to PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
  /export/pdf/subject/{id}:
    get:
      operationId: exportSubjectPdf
      tags: [export]
      summary: Export all questions in a subject to PDF
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
  /export/pdf/selected:
    post:
      operationId: exportSelectedPdf
      tags: [export]
      summary: Export selected questions to PDF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportSelectedBody"
      responses:
        "200":
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Chapter:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        subjectName:
          type: ["string", "null"]
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        questionCount:
          type: ["integer", "null"]
      required:
        - id
        - subjectId
        - name
        - createdAt
    CreateChapterBody:
      type: object
      properties:
        subjectId:
          type: integer
        name:
          type: string
      required:
        - subjectId
        - name
...
    Choice:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        text:
          type: string
        imageUrl:
          type: ["string", "null"]
        imageName:
          type: ["string", "null"]
        isCorrect:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - questionId
        - text
        - isCorrect
        - createdAt
    CreateChoiceBody:
      type: object
      properties:
        questionId:
          type: integer
        text:
          type: string
        isCorrect:
          type: boolean
        image:
          type: string
          format: binary
      required:
        - questionId
        - text
        - isCorrect
    UpdateChoiceBody:
      type: object
      properties:
        text:
          type: string
        isCorrect:
          type: boolean
        image:
          type: string
          format: binary
        removeImage:
          type: string
          enum: ["true", "false"]
...
    ExportSelectedBody:
      type: object
      properties:
        questionIds:
          type: array
          items:
            type: integer
        title:
          type: string
      required:
        - questionIds