# lib/api-zod/src/generated/types/createChapterBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Question Bank API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateChapterBody {
  subjectId: number;
  name: string;
}